ml/requirements.txt
tensorflow>=2.13.0
numpy>=1.23.5
matplotlib>=3.7.1
scikit-learn>=1.3.0
tensorflowjs>=4.10.0
Pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.3
